[icloud] Update AccountHandler to prevent null pointers from crashing binding.
I recently ran into an exception which crashed the iCloud binding:
2019-09-22 21:33:11.322 [WARN ] [mmon.WrappedScheduledExecutorService] - Scheduled runnable ended with an exception:
java.lang.NullPointerException: null
        at org.openhab.binding.icloud.internal.handler.ICloudAccountBridgeHandler.refreshData(ICloudAccountBridgeHandler.java:154) ~[?:?]
        at java.util.concurrent.Executors$RunnableAdapter.call(Executors.java:511) ~[?:?]
        at java.util.concurrent.FutureTask.runAndReset(FutureTask.java:308) ~[?:?]
        at java.util.concurrent.ScheduledThreadPoolExecutor$ScheduledFutureTask.access$301(ScheduledThreadPoolExecutor.java:180) ~[?:?]
        at java.util.concurrent.ScheduledThreadPoolExecutor$ScheduledFutureTask.run(ScheduledThreadPoolExecutor.java:294) ~[?:?]
        at java.util.concurrent.ThreadPoolExecutor.runWorker(ThreadPoolExecutor.java:1149) [?:?]
        at java.util.concurrent.ThreadPoolExecutor$Worker.run(ThreadPoolExecutor.java:624) [?:?]
        at java.lang.Thread.run(Thread.java:748) [?:?]

It seems like the NPE happened because iCloudData from
iCloudData.getICloudAccountStatusCode() was null. Following the stack
up, iCloudData is created from a GSON.fromGson(...) call, which can
actually return null if the json is empty! I assume this can happen if
there is a problem with the server response.

To prevent this from happening, we can move the null check after GSON
parses the string.

Signed-off-by: Maury Quijada <maury.quijada@gmail.com>